+++ layer/src/main/scala/geotrellis/layer/merge/Mergable.scala
@@ -1,5 +1,0 @@
-package geotrellis.layer.merge
-
-trait Mergable[T] {
-  def merge(t1: T, t2: T): T
-}