+++ spark/src/main/scala/geotrellis/spark/store/LayerWriter.scala
@@ -21,7 +21,6 @@ import geotrellis.store._
 import geotrellis.store.avro._
 import geotrellis.store.avro.codecs._
 import geotrellis.store.index._
-import geotrellis.layer.merge.Mergable
 import geotrellis.spark._
 import geotrellis.util._
 
@@ -29,11 +28,11 @@ import com.typesafe.scalalogging.LazyLogging
 import org.apache.avro._
 import org.apache.spark.rdd.RDD
 import _root_.io.circe._
+import cats.Semigroup
 
 import scala.reflect.ClassTag
 import java.util.ServiceLoader
 import java.net.URI
-
 
 trait LayerWriter[ID] {
   val attributeStore: AttributeStore
@@ -48,7 +47,7 @@ trait LayerWriter[ID] {
     H: Encoder: Decoder,
     K: AvroRecordCodec: Boundable: Encoder: ClassTag,
     V: AvroRecordCodec,
-    M: Component[?, Bounds[K]]: Mergable: Encoder: Decoder
+    M: Component[?, Bounds[K]]: Semigroup: Encoder: Decoder
   ](id: LayerId, updateMetadata: M): Option[LayerAttributes[H, M, K]] = {
     if (!attributeStore.layerExists(id)) throw new LayerNotFoundError(id)
 
@@ -93,7 +92,7 @@ trait LayerWriter[ID] {
   def update[
     K: AvroRecordCodec: Boundable: Encoder: Decoder: ClassTag,
     V: AvroRecordCodec: ClassTag,
-    M: Encoder: Decoder: Component[?, Bounds[K]]: Mergable
+    M: Encoder: Decoder: Component[?, Bounds[K]]: Semigroup
   ](id: ID, rdd: RDD[(K, V)] with Metadata[M], mergeFunc: (V, V) => V = { (existing: V, updating: V) => updating }): Unit
 
   /** Update persisted layer without checking for possible.
@@ -118,7 +117,7 @@ trait LayerWriter[ID] {
   def overwrite[
     K: AvroRecordCodec: Boundable: Encoder: Decoder: ClassTag,
     V: AvroRecordCodec: ClassTag,
-    M: Encoder: Decoder: Component[?, Bounds[K]]: Mergable
+    M: Encoder: Decoder: Component[?, Bounds[K]]: Semigroup
   ](id: ID, rdd: RDD[(K, V)] with Metadata[M]): Unit
 
   // Layer Writing